Finalize course

+++ Shop/Data/Models/ShopCartItem.cs
@@ -9,7 +9,7 @@ namespace Shop.Data.Models
     {
         public int id { get; set; }
         public Car car { get; set; }
-        public int price { get; set; }
+        public ushort price { get; set; }
         
         public string ShopCartId { get; set; }
     }
